Publish-ja: Change theme path from 17bad17

+++ Makefile
@@ -7,7 +7,7 @@ PUBLISH_PAGES_DIR=.publish
 EMACS_DEPS_DIR=.emacs-dependencies
 DIST_DIR=dist
 
-install: setup-git get-deps
+install: setup-git mkdirs get-deps
 
 setup-git:
 	git submodule update
@@ -19,9 +19,14 @@ install-emacs:
 	sudo apt update
 	sudo apt-get install -y emacs25
 
+mkdirs:
+	if [ ! -d "$(EMACS_DEPS_DIR)" ]; then mkdir -p "$(EMACS_DEPS_DIR)"; fi
+	if [ ! -d "$(GH_PAGES_DIR)" ]; then mkdir -p "$(GH_PAGES_DIR)"; fi
+	if [ ! -d "$(PUBLISH_PAGES_DIR)" ]; then mkdir -p "$(PUBLISH_PAGES_DIR)"; fi
+	if [ ! -d "$(DIST_DIR)" ]; then mkdir -p "$(DIST_DIR)"; fi
+
 get-deps:
 	sudo apt-get install -y texinfo
-	mkdir "$(EMACS_DEPS_DIR)"
 	git clone https://github.com/hniksic/emacs-htmlize.git "$(EMACS_DEPS_DIR)/emacs-htmlize"
 	git clone https://github.com/mrlee23/org-multilingual.git "$(EMACS_DEPS_DIR)/org-multilingual"
 	git clone -b release_9.0.10 https://code.orgmode.org/bzg/org-mode.git "$(EMACS_DEPS_DIR)/org-mode"
@@ -31,23 +36,18 @@ test:
 	echo "No tests."
 	echo "HI"
 
-pages:
-	rm -rf .org-timestamps
+pages: mkdirs pages-deps
+	if [ -d .org-timestamps ]; then rm -rf .org-timestamps; fi
+	if [ -f "CNAME" ]; then cp CNAME "$(GH_PAGES_DIR)/CNAME"; fi
+	if [ -d "$(DIST_DIR)" ]; then mv -f "$(DIST_DIR)" "$(GH_PAGES_DIR)/$(DIST_DIR)"; fi
 	@$(BATCH) --eval "(progn\
 	$$load_paths\
 	(message \"%s\" load-path)\
 	(load-file \"pages.el\")\
 	(pages-publish \"./\" \"$(GH_PAGES_DIR)/\" \"$(PUBLISH_PAGES_DIR)\")\
 	)"
-	if [ -f "CNAME" ]; then cp CNAME "$(GH_PAGES_DIR)/CNAME"; fi
-	# rm -rf "$(GH_PAGES_DIR)/$(EMACS_DEPS_DIR)"
-	# rm -rf "$(EMACS_DEPS_DIR)"
 	for lang in `cat .supportLanguages`; \
 	do \
-		if [ -d "$(DIST_DIR)" ]; \
-		then \
-			cp -rf "$(DIST_DIR)" "$(GH_PAGES_DIR)/$${lang}/$(DIST_DIR)"; \
-		fi; \
 		echo "$$lang" > "$(PUBLISH_PAGES_DIR)/$${lang}/.language"; \
 	done;
 	find . -d -path "*/$(DIST_DIR)/*/.git" -prune -exec rm -rf {} \;
